Rename to server to servertools
