--- server/context_key.go
@@ -1,8 +1,0 @@
-package server
-
-type ContextKey string
-
-var (
-	ContextKeySystemId   = ContextKey("systemId")
-	ContextKeySystemType = ContextKey("systemType")
-)
