+++ signin/go.mod
@@ -5,7 +5,7 @@ go 1.16
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/gorilla/mux v1.8.0
-	github.com/weekendprojectapp/authful/server v0.0.0
+	github.com/weekendprojectapp/authful/servertools v0.0.0
 )
 
-replace github.com/weekendprojectapp/authful/server => ../server
+replace github.com/weekendprojectapp/authful/servertools => ../servertools
