+++ signin/internal/web/web.go
@@ -6,7 +6,7 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/weekendprojectapp/authful/server"
+	"github.com/weekendprojectapp/authful/servertools"
 	"github.com/weekendprojectapp/authful/signin/pkg/models"
 )
 
@@ -29,12 +29,12 @@ func AuthorizeUser(w http.ResponseWriter, r *http.Request) {
 	var loginRequest models.SigninCredentials
 	json.NewDecoder(r.Body).Decode(&loginRequest)
 
-	server.HandleResponse(w, []byte{}, http.StatusUnauthorized)
+	servertools.HandleResponse(w, []byte{}, http.StatusUnauthorized)
 
 }
 
 func GetProfile(w http.ResponseWriter, r *http.Request) {
-	server.HandleResponse(w, []byte{}, http.StatusOK)
+	servertools.HandleResponse(w, []byte{}, http.StatusOK)
 }
 
 func Index(w http.ResponseWriter, r *http.Request) {
