+++ signin/main.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/dgrijalva/jwt-go"
-	srvr "github.com/weekendprojectapp/authful/server"
+	"github.com/weekendprojectapp/authful/servertools"
 	"github.com/weekendprojectapp/authful/signin/internal/config"
 	"github.com/weekendprojectapp/authful/signin/internal/web"
 
@@ -91,9 +91,9 @@ func processToken(rawToken string, r *http.Request) (bool, *http.Request) {
 	systemType := ""
 	isValid := false
 
-	var claims srvr.Claims
+	var claims servertools.Claims
 	token, err := jwt.ParseWithClaims(rawToken, &claims, func(t *jwt.Token) (interface{}, error) {
-		localClaim := t.Claims.(*srvr.Claims)
+		localClaim := t.Claims.(*servertools.Claims)
 		systemId = localClaim.SystemId
 		systemType = localClaim.Type
 		return []byte(config.GetConfig().Security.JwtKey), nil
@@ -107,8 +107,8 @@ func processToken(rawToken string, r *http.Request) (bool, *http.Request) {
 		fmt.Println("Error happened: " + err.Error())
 	}
 
-	ctx := context.WithValue(r.Context(), srvr.ContextKeySystemId, systemId)
-	ctx = context.WithValue(ctx, srvr.ContextKeySystemType, systemType)
+	ctx := context.WithValue(r.Context(), servertools.ContextKeySystemId, systemId)
+	ctx = context.WithValue(ctx, servertools.ContextKeySystemType, systemType)
 	r = r.WithContext(ctx)
 
 	return isValid, r
